feat!: land the editorial layer as zingolib::perspective, not a crate
The zingo-perspective extraction (PR #2611) is declined before merge:
the funnel's repoint churn across three consumer repositories outweighs
the gate maintenance the in-crate form costs. The editorial layer — the
value-transfer derivation, the finsight rollups, and their types —
moves instead to a zingolib::perspective module behind the crate's
first default-on feature, whose off build CI's feature-powerset job
compiles as a standing proof that core stands without its editorial
layer. Rule 7 of ADR 0024 remains as originally written, and this
commit's amendment records the reversal, the asymmetric boundary, and
the accepted costs.

The derivation returns to inherent impl blocks, so no call site
changes; consumers rewrite only the use lines naming the moved types.
The eight golden fixtures port from the seed branch and reproduce
byte-identically, pinning the pre-extraction contract.
