+++ libtonode-tests/tests/concrete.rs
@@ -548,12 +548,12 @@ use zcash_local_net::validator::Validator;
 use zingolib::config::{ChainType, ClientConfig};
 use zingolib::lightclient::LightClient;
 use zingolib::lightclient::error::{LightClientError, SendError};
+use zingolib::perspective::value_transfer::{
+    SelfSendValueTransfer, SentValueTransfer, ValueTransferKind,
+};
 use zingolib::testutils::build_fvks_from_unified_keystore;
 use zingolib::wallet::error::CalculateTransactionError;
 use zingolib::wallet::output::SpendStatus;
-use zingolib::wallet::summary::data::{
-    SelfSendValueTransfer, SentValueTransfer, ValueTransferKind,
-};
 
 #[tokio::test]
 async fn test_scanning_in_watch_only_mode() {
