+++ zingolib/src/testutils/chain_generics/fixtures.rs
@@ -5,16 +5,18 @@ use zcash_primitives::transaction::fees::zip317::MARGINAL_FEE;
 use zcash_protocol::value::Zatoshis;
 use zcash_protocol::{PoolType, ShieldedPool};
 
+#[cfg(feature = "perspective")]
+use crate::perspective::value_transfer::{
+    SelfSendValueTransfer, SentValueTransfer, ValueTransferKind,
+};
 use crate::testutils::chain_generics::conduct_chain::ConductChain;
 use crate::testutils::chain_generics::with_assertions;
 use crate::testutils::fee_tables;
 use crate::testutils::lightclient::get_base_address;
 use crate::testutils::timestamped_test_log;
-use crate::wallet::summary::data::SelfSendValueTransfer;
-use crate::wallet::summary::data::SentValueTransfer;
-use crate::wallet::summary::data::ValueTransferKind;
 
 /// Fixture for testing various vt transactions
+#[cfg(feature = "perspective")]
 pub async fn create_various_value_transfers<CC>()
 where
     CC: ConductChain,
