+++ zingolib/tests/golden/messages_all.json
@@ -1,0 +1,76 @@
+{
+  "value_transfers": [
+    {
+      "txid": "0101010101010101010101010101010101010101010101010101010101010101",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 10,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "received",
+      "value": 1600000,
+      "recipient_address": null,
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": [
+        "Hello Alice"
+      ]
+    },
+    {
+      "txid": "0303030303030303030303030303030303030303030303030303030303030303",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 12,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "sent",
+      "value": 800000,
+      "recipient_address": "uregtest155cx8wdxvqfulfysc9nzw85s8rq2fceepar77lsgm7h68ramkntxdk6pmjctnqtazv5hd40my0zp8555lhd2swddzxy8h8gjw64rrv43kuslg05vfsep36y8xkhtaz6ytwkqg0uzgughjdu59v67esqsek7p3aahxu0zmjdkdcp3dewk",
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": [
+        "hi Bob"
+      ]
+    },
+    {
+      "txid": "0404040404040404040404040404040404040404040404040404040404040404",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 13,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "memo-to-self",
+      "value": 800000,
+      "recipient_address": null,
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": [
+        "note to my future self"
+      ]
+    },
+    {
+      "txid": "0404040404040404040404040404040404040404040404040404040404040404",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 13,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "sent",
+      "value": 800000,
+      "recipient_address": "uregtest155cx8wdxvqfulfysc9nzw85s8rq2fceepar77lsgm7h68ramkntxdk6pmjctnqtazv5hd40my0zp8555lhd2swddzxy8h8gjw64rrv43kuslg05vfsep36y8xkhtaz6ytwkqg0uzgughjdu59v67esqsek7p3aahxu0zmjdkdcp3dewk",
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": [
+        "second memo to Bob"
+      ]
+    }
+  ]
+}
