+++ zingolib/tests/golden/total_spends_to_address.json
@@ -1,0 +1,4 @@
+{
+  "uregtest14emvr2anyul683p43d0ck55c04r65ld6f0shetcn77z8j7m64hm4ku3wguf60s75f0g3s7r7g89z22f3ff5tsfgr45efj4pe2gyg5krqp5vvl3afu0280zp9ru2379zat5y6nkqkwjxsvpq5900kchcgzaw8v8z3ggt5yymnuj9hymtv3p533fcrk2wnj48g5vg42vle08c2xtanq0e": 1,
+  "uregtest155cx8wdxvqfulfysc9nzw85s8rq2fceepar77lsgm7h68ramkntxdk6pmjctnqtazv5hd40my0zp8555lhd2swddzxy8h8gjw64rrv43kuslg05vfsep36y8xkhtaz6ytwkqg0uzgughjdu59v67esqsek7p3aahxu0zmjdkdcp3dewk": 2
+}
