+++ zingolib/tests/golden/value_transfers.json
@@ -1,0 +1,138 @@
+{
+  "value_transfers": [
+    {
+      "txid": "0606060606060606060606060606060606060606060606060606060606060606",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 15,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "send-to-self",
+      "value": 0,
+      "recipient_address": null,
+      "pools_sent_from": [],
+      "pools_received": [],
+      "memos": []
+    },
+    {
+      "txid": "0606060606060606060606060606060606060606060606060606060606060606",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 15,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "sent",
+      "value": 800000,
+      "recipient_address": "uregtest14emvr2anyul683p43d0ck55c04r65ld6f0shetcn77z8j7m64hm4ku3wguf60s75f0g3s7r7g89z22f3ff5tsfgr45efj4pe2gyg5krqp5vvl3afu0280zp9ru2379zat5y6nkqkwjxsvpq5900kchcgzaw8v8z3ggt5yymnuj9hymtv3p533fcrk2wnj48g5vg42vle08c2xtanq0e",
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": []
+    },
+    {
+      "txid": "0505050505050505050505050505050505050505050505050505050505050505",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 14,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "send-to-self",
+      "value": 800000,
+      "recipient_address": null,
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": []
+    },
+    {
+      "txid": "0404040404040404040404040404040404040404040404040404040404040404",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 13,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "sent",
+      "value": 800000,
+      "recipient_address": "uregtest155cx8wdxvqfulfysc9nzw85s8rq2fceepar77lsgm7h68ramkntxdk6pmjctnqtazv5hd40my0zp8555lhd2swddzxy8h8gjw64rrv43kuslg05vfsep36y8xkhtaz6ytwkqg0uzgughjdu59v67esqsek7p3aahxu0zmjdkdcp3dewk",
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": [
+        "second memo to Bob"
+      ]
+    },
+    {
+      "txid": "0404040404040404040404040404040404040404040404040404040404040404",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 13,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "memo-to-self",
+      "value": 800000,
+      "recipient_address": null,
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": [
+        "note to my future self"
+      ]
+    },
+    {
+      "txid": "0303030303030303030303030303030303030303030303030303030303030303",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 12,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "sent",
+      "value": 800000,
+      "recipient_address": "uregtest155cx8wdxvqfulfysc9nzw85s8rq2fceepar77lsgm7h68ramkntxdk6pmjctnqtazv5hd40my0zp8555lhd2swddzxy8h8gjw64rrv43kuslg05vfsep36y8xkhtaz6ytwkqg0uzgughjdu59v67esqsek7p3aahxu0zmjdkdcp3dewk",
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": [
+        "hi Bob"
+      ]
+    },
+    {
+      "txid": "0202020202020202020202020202020202020202020202020202020202020202",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 11,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "received",
+      "value": 800000,
+      "recipient_address": null,
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": []
+    },
+    {
+      "txid": "0101010101010101010101010101010101010101010101010101010101010101",
+      "datetime": 0,
+      "status": "confirmed",
+      "blockheight": 10,
+      "transaction_fee": 0,
+      "zec_price": null,
+      "kind": "received",
+      "value": 1600000,
+      "recipient_address": null,
+      "pools_sent_from": [],
+      "pools_received": [
+        "Ironwood"
+      ],
+      "memos": [
+        "Hello Alice"
+      ]
+    }
+  ]
+}
